llm_envs: collapse HuggingFaceGym/IterablePromptBatchGym into DatasetEnv; drop ReasoningGym + make_reasoning_rollout_env
- Fold HuggingFaceGym + IterablePromptBatchGym into one concrete
  DatasetEnv(LLMEnv, gym.Env); PreferenceGym/SFTGym stay thin shims.
- Delete ReasoningGym + its equivalence oracle; keep the standalone
  RolloutEnv/BatchIterationState property tests.
- Remove make_reasoning_rollout_env; its 4 example/test callers build
  RolloutEnv(max_turns=1) + TokenObservationWrapper directly.
- train_llm hints + isinstance tests repoint to DatasetEnv.
- Fix BatchRolloutEnv.reset to guard dataset_size via getattr and only pass
  row_index when resolved (regression from the BatchIterationState move).

Net ~-1700 lines; 423 llm-env/protocol/train/util/rollout tests pass.
