+++ agilerl/llm_envs/__init__.py
@@ -2,18 +2,16 @@
 
 from agilerl.llm_envs import search as _search
 from agilerl.llm_envs.base import (
-    HuggingFaceGym,
-    IterablePromptBatchGym,
     LLMEnv,
     apply_chat_template,
 )
 from agilerl.llm_envs.dataset_env import DatasetEnv, PreferenceGym, SFTGym
-from agilerl.llm_envs.reasoning import ReasoningGym
 from agilerl.llm_envs.rollout_env import (
     BatchIterationState,
     RolloutEnv,
+    _default_prompt_builder,
+    _extract_question_answer_columns,
     dataloader_shuffle_order,
-    make_reasoning_rollout_env,
 )
 from agilerl.llm_envs.search import TIMEOUT, FormatRewardWrapper, SearchTool
 from agilerl.llm_envs.sync_vec_env import (
@@ -32,11 +30,8 @@ __all__ = [
     "BatchRolloutEnv",
     "DatasetEnv",
     "FormatRewardWrapper",
-    "HuggingFaceGym",
-    "IterablePromptBatchGym",
     "LLMEnv",
     "PreferenceGym",
-    "ReasoningGym",
     "RolloutEnv",
     "SFTGym",
     "SearchTool",
@@ -44,8 +39,9 @@ __all__ = [
     "TokenObservationWrapper",
     "Trajectory",
     "TrajectoryBuffer",
+    "_default_prompt_builder",
+    "_extract_question_answer_columns",
     "apply_chat_template",
     "dataloader_shuffle_order",
-    "make_reasoning_rollout_env",
     "requests",
 ]
