+++ agilerl/llm_envs/dataset_env.py
@@ -10,14 +10,21 @@ one class. ``PreferenceGym`` / ``SFTGym`` remain as thin back-compat subclasses.
 
 from __future__ import annotations
 
+import copy
+import warnings
+from collections.abc import Callable
+from contextlib import contextmanager
 from typing import TYPE_CHECKING, Any
 
-from agilerl.llm_envs.base import IterablePromptBatchGym
+import gymnasium as gym
+import torch
+from torch.utils.data import DataLoader
+
+from agilerl.llm_envs.base import LLMEnv
 
 if TYPE_CHECKING:
-    from collections.abc import Callable
+    from collections.abc import Generator
 
-    import torch
     from accelerate import Accelerator
     from datasets import Dataset
     from transformers import AutoTokenizer
@@ -28,13 +35,17 @@ if TYPE_CHECKING:
     ]
 
 
-class DatasetEnv(IterablePromptBatchGym):
+class DatasetEnv(LLMEnv, gym.Env):
     """Teacher-forced, dataset-backed LLM env (no generation).
 
-    Configured by ``required_columns`` + a ``collate_builder`` (and optional
-    ``response_column``) instead of a subclass per training regime. ``reset``/``step`` (the
-    dataloader advance, completions ignored) are inherited from
-    :class:`~agilerl.llm_envs.base.IterablePromptBatchGym`.
+    The no-generation half of the env taxonomy: completions are dataset labels
+    scored in a single teacher-forced forward (SFT cross-entropy, DPO preference)
+    with no autoregressive rollout. The training regimes (preference / SFT) differ
+    only by the *required columns* and the *collate function* — descriptors, not
+    subclasses — configured via ``required_columns`` + a ``collate_builder`` (and
+    optional ``response_column``). ``reset`` / ``step`` advance the seeded
+    ``DataLoader`` (completions ignored); ``PreferenceGym`` / ``SFTGym`` remain as
+    thin back-compat subclasses.
     """
 
     def __init__(
@@ -60,17 +71,54 @@ class DatasetEnv(IterablePromptBatchGym):
             assert self.required_columns.issubset(set(dataset.features.keys())), (
                 f"{label} dataset must contain columns {sorted(self.required_columns)}."
             )
-        super().__init__(
-            train_dataset=train_dataset,
-            test_dataset=test_dataset,
-            tokenizer=tokenizer,
-            conversation_template=None,
-            data_batch_size_per_gpu=data_batch_size_per_gpu,
-            max_context_length=max_context_length,
-            min_completion_length=min_completion_length,
-            accelerator=accelerator,
-            seed=seed,
+
+        self.name = train_dataset.info.dataset_name
+        self.tokenizer = tokenizer
+        self.data_batch_size_per_gpu = data_batch_size_per_gpu
+        self.accelerator = accelerator
+        self.min_completion_length = (
+            0 if min_completion_length is None else min_completion_length
         )
+        self.max_context_length = max_context_length
+        self.seed = seed
+        generator = torch.Generator().manual_seed(seed)
+        custom_collate_fn = self.create_collate_fn(tokenizer)
+        dataloader_kwargs = {"collate_fn": custom_collate_fn}
+        train_dataset = self._filter_dataset_by_max_context_length(
+            train_dataset,
+            "train dataset",
+        )
+        test_dataset = self._filter_dataset_by_max_context_length(
+            test_dataset,
+            "test dataset",
+        )
+        self.train_dataloader = DataLoader(
+            train_dataset,
+            batch_size=data_batch_size_per_gpu,
+            shuffle=True,
+            **dataloader_kwargs,
+            generator=generator,
+        )
+        self.test_dataloader = DataLoader(
+            test_dataset,
+            batch_size=data_batch_size_per_gpu,
+            shuffle=False,
+            **dataloader_kwargs,
+            generator=generator,
+        )
+        self.dataset_size = {
+            "train": len(train_dataset),
+            "test": len(test_dataset),
+        }
+        if self.accelerator is not None:
+            self.train_dataloader = self.accelerator.prepare(self.train_dataloader)
+            self.test_dataloader = self.accelerator.prepare(self.test_dataloader)
+        self.train_dataloader_iter = iter(self.train_dataloader)
+        self.test_dataloader_iter = iter(self.test_dataloader)
+        self.dataloader = self.train_dataloader_iter
+        self.reset_called = False
+        self.evaluation_mode = False
+        self.num_epochs = 0
 
     def create_collate_fn(
         self,
@@ -79,6 +127,107 @@ class DatasetEnv(IterablePromptBatchGym):
     ) -> Callable[[list[dict[str, Any]]], dict[str, Any]]:
         """Delegate to the injected collate builder."""
         return self._collate_builder(self, tokenizer, max_context_length)
+
+    def reset(self, reset_dataloaders: bool = False) -> Any:
+        """Reset the environment and get the next batch from the dataloader."""
+        if reset_dataloaders:
+            self._reset_dataloaders()
+            warnings.warn(
+                "env.reset() called with reset_dataloaders=True, this will reset "
+                "the dataloaders to the beginning of the dataset, proceed with caution.",
+                stacklevel=2,
+            )
+        if self.reset_called:
+            warnings.warn(
+                "env.reset() called more than once sequentially, it should typically "
+                "follow with env.step().",
+                stacklevel=2,
+            )
+        self.reset_called = True
+        return self._get_next_batch()
+
+    def step(self, completions: torch.Tensor | None = None) -> Any:
+        """Advance the iterator and return the next batch (``completions`` unused)."""
+        self.reset_called = False
+        return self._get_next_batch()
+
+    def _get_next_batch(self) -> Any:
+        try:
+            batch = next(self.dataloader)
+        except StopIteration:
+            if not self.evaluation_mode:
+                self.num_epochs += 1
+            self._reset_dataloaders(
+                reset_train=not self.evaluation_mode,
+                reset_test=self.evaluation_mode,
+            )
+            return self._get_next_batch()
+        return batch
+
+    @contextmanager
+    def eval_mode(self) -> Generator[None, None, None]:
+        """Context manager to switch to evaluation mode."""
+        self.dataloader = self.test_dataloader_iter
+        self.evaluation_mode = True
+        last_tokenized_prompts = None
+        if hasattr(self, "last_tokenized_prompts"):
+            last_tokenized_prompts = copy.deepcopy(self.last_tokenized_prompts)
+        try:
+            yield
+        finally:
+            self.dataloader = self.train_dataloader_iter
+            self.evaluation_mode = False
+            if last_tokenized_prompts is not None:
+                self.last_tokenized_prompts = last_tokenized_prompts
+
+    def __len__(self) -> int:
+        """Return the length of the dataset."""
+        if self.evaluation_mode:
+            return len(self.test_dataloader.dataset)
+        return len(self.train_dataloader.dataset)
+
+    def _reset_dataloaders(
+        self, reset_train: bool = True, reset_test: bool = True
+    ) -> None:
+        """Reset the dataloaders to the beginning of the dataset."""
+        if reset_train:
+            self.train_dataloader_iter = iter(self.train_dataloader)
+        if reset_test:
+            self.test_dataloader_iter = iter(self.test_dataloader)
+        self.dataloader = (
+            self.test_dataloader_iter
+            if self.evaluation_mode
+            else self.train_dataloader_iter
+        )
+
+    def _filter_dataset_by_max_context_length(
+        self,
+        dataset: Dataset,
+        dataset_type: str | None = None,
+    ) -> Dataset:
+        """Filter the dataset by the max context length."""
+        dataset_type = "dataset" if dataset_type is None else dataset_type
+        filter_keyword = "prompt" if "prompt" in dataset.features else "question"
+        if self.max_context_length is None or not isinstance(
+            dataset[0][filter_keyword],
+            str,
+        ):
+            return dataset
+        filtered_dataset = dataset.filter(
+            lambda x: (
+                len(self.tokenizer.encode(x[filter_keyword]))
+                <= self.max_context_length - self.min_completion_length
+            ),
+        )
+        if len(filtered_dataset) == 0:
+            msg = f"No samples left in the {dataset_type} after filtering by the max context length constraint, use a larger max context length."
+            raise ValueError(msg)
+        if (dataset_difference := len(dataset) - len(filtered_dataset)) > 0:
+            warnings.warn(
+                f"{dataset_difference} samples were filtered out of the {dataset_type} due to the max context length constraint.",
+                stacklevel=2,
+            )
+        return filtered_dataset
 
 
 def preference_collate_builder(
