+++ docs/api/wrappers/llm_envs.rst
@@ -5,8 +5,8 @@ Gymnasium-style environments for supervised fine-tuning, preference optimization
 and reasoning RL. These types are also re-exported from
 :mod:`agilerl.utils.llm_utils` for backwards compatibility.
 
-.. autoclass:: agilerl.llm_envs.HuggingFaceGym
+.. autoclass:: agilerl.llm_envs.DatasetEnv
 .. autoclass:: agilerl.llm_envs.SFTGym
 .. autoclass:: agilerl.llm_envs.PreferenceGym
-.. autoclass:: agilerl.llm_envs.ReasoningGym
+.. autoclass:: agilerl.llm_envs.RolloutEnv
 .. autofunction:: agilerl.llm_envs.apply_chat_template
