+++ tutorials/llm_finetuning/grpo_reasoning.py
@@ -9,7 +9,12 @@ from transformers import AutoTokenizer
 from agilerl.algorithms import GRPO
 from agilerl.training.train_llm import finetune_llm_multiturn
 from agilerl.utils.algo_utils import VLLMConfig
-from agilerl.llm_envs import make_reasoning_rollout_env
+from agilerl.llm_envs import (
+    RolloutEnv,
+    TokenObservationWrapper,
+    _default_prompt_builder,
+    _extract_question_answer_columns,
+)
 
 MODEL_PATH = "Qwen/Qwen2.5-0.5B"
 DATASET = "Jiayi-Pan/Countdown-Tasks-3to4"
@@ -124,16 +129,28 @@ def main():
     # dataset cursor, giving each GRPO group a deterministic, group-consistent
     # dataset order.
     def env_factory(evaluation_mode: bool = False):
-        return make_reasoning_rollout_env(
-            train_dataset=train_dataset,
-            test_dataset=test_dataset,
-            tokenizer=tokenizer,
+        train_questions, train_answers = _extract_question_answer_columns(train_dataset)
+        test_questions, test_answers = _extract_question_answer_columns(test_dataset)
+        raw_env = RolloutEnv(
+            max_turns=1,
+            questions=train_questions,
+            answers=train_answers,
             reward_fn=combined_rewards,
-            conversation_template=conversation_template,
-            evaluation_mode=evaluation_mode,
-            seed=42,
+            prompt_builder=_default_prompt_builder(conversation_template),
+            test_questions=test_questions,
+            test_answers=test_answers,
+        )
+        raw_env.evaluation_mode = evaluation_mode
+        wrapper = TokenObservationWrapper(
+            raw_env,
+            tokenizer=tokenizer,
+            max_turns=1,
+            pad_id=getattr(tokenizer, "pad_token_id", None),
+            apply_chat_template=True,
             max_model_len=MAX_CONTEXT_LENGTH,
         )
+        wrapper.eval_mode = raw_env.eval_mode
+        return wrapper
 
     # Define the LoRA configuration
     lora_config = LoraConfig(
